change texture filter to nearest set run scene
The default godot texture filter is Linear but
it makes the pixel art looks blurry. To avoid
that use the option Project -> Project Settings
-> Texture -> Default Texture Filter and set
it to Nearest.

When godot game is run for the first time it
asks for the first scene to be executed. We
just select the current one (Game).