add the player scene to game and create a camera
Add some frames from the knight.png spritesheet
to the player's AnimatedSprite2D, change the
fps to 10 and move the sprite over the red
base line.

Add a CollisionShape2D and set the Shape as a
cirle that should be inside the area covered by
the animated sprite. Important that it touches
the base red line too so the character would not
float over or above the ground.

Save the player scene.

Add the player scene to the game scene by simply
dragging the player.tscn file from assets to the
scene panel of the game.