add coin collision with player
We want coin _on_body_entered() to detect
collision with the player only so we put
the collision level of the player and the
collision mask of the coin both on 2
deselecting the default 1.
We then add a en event handler to the Coin
body_entered signal (Coin's Node panel)
and temporary add a debug print to it.